refactor(BE): recreate seeds for better code structure
